work2_1/requirements.txt
addict
future
lmdb
numpy>=1.17
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch>=1.7
torchvision
pytorch-msssim
lightning
lpips
tensorboardx
tqdm
yapf
timm
einops
kornia
matplotlib
yacs
joblib
natsort
h5py
tqdm